refactor(cli): move template platform validation to better spot
